Merge pull request #155 from depromeet/staging

+++ src/main/java/kr/co/antoon/oauth/application/AuthService.java
@@ -54,16 +54,17 @@ public class AuthService {
     }
 
     @Transactional
-    public void revokeToken(String accesss, String refresh) {
-        var access = accesss.substring(7);
+    public void revokeToken(String accessToken) {
+        var access = accessToken.substring(7);
 
         if (!jwtTokenProvider.validate(access)) { //유효성 검사
             throw new TokenExpiredException(ErrorMessage.NOT_VALIDATE_TOKEN);
         }
 
-        var userId = jwtTokenProvider.getUserId(access);
-        var user = userRepository.findById(userId)
-                .orElseThrow(() -> new NotExistsException(ErrorMessage.NOT_EXIST_USER));
+        Long userId = jwtTokenProvider.getUserId(access);
+        if(!userRepository.existsById(userId)) {
+            throw new NotExistsException(ErrorMessage.NOT_EXIST_USER);
+        }
 
         if (userRedisCacheService.get(userId) != null) {
             userRedisCacheService.delete(userId);
